pyproject.toml
[project]
name = "music-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=1.2.3",
    "langchain-xai>=1.2.1",
    "python-telegram-bot>=22.5",
    "ytmusicapi>=1.11.4",
]
